packages/litqa/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=64", "setuptools_scm>=8"]

[project]
authors = [
    {email = "hello@futurehouse.org", name = "FutureHouse technical staff"},
]
classifiers = [
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python",
]
dependencies = [
    "datasets",
    "fh-llm-client>=0.0.11",
    "ldp>=0.20",
    "paper-qa @ https://github.com/Future-House/paper-qa.git",
    "pydantic~=2.0",
    "tenacity",
]
description = "LitQA environment implemented with aviary"
dynamic = ["version"]
name = "aviary.litqa"
readme = "README.md"
requires-python = ">=3.11"

[project.optional-dependencies]
datasets = ["datasets"]

[tool.ruff]
extend = "../../pyproject.toml"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
root = "../.."
version_file = "src/aviary/envs/litqa/version.py"
